Use a flux-form formulation for Saint-Venant (rather than tendencies)
Ignore-this: 510f3aad72bea416bc94f6ca684ba831

The main reason for this change is that the flux-form formulation
parallelises easily, which is not the case for the previous
"tendencies" formulation which requires concurrent access to
tendencies. This is unfortunate since the tendencies formulation is
significantly simpler (in serial) and uses less memory.

darcs-hash:e226a28c544b13102ee8774d73f0928baf599c5e